Add files via upload (#54)
For Fizz Buzz, I used the modulus operator to determine if the number was a multiple of 3 or 5. I then changed the output based on that determination.

For String calculator, I split the input string into each character. For each character, I determined if the value could be changed to a number. If it was possible, I added that number to a total and returned the total.