Statler Smith HW3
I evaluated the Big O functions based on the inputs and the processing.

For persistence, I used multiplications of 10 to separate out the input number into the digits. I then multiplied the digits together to generate a new number that I performed the same processing on. I did this until the new number was only one digit. I recorded the number of iterations I went through and returned that result.