Hoyong Jung hw3 (#77)
* Solve fizz-buzz and stringCalculator problems

This commit will add the solutions to fizz-buzz and stingCalculator problems. For fizz-buzz, it will print 'Fizz' when num % 3 === 0, 'Buzz' when num % 5 === 0, and 'Fizz-Buzz' when num % 15 === 0. For StringCalculator, I split the string with comma then loop through the array to add the numbers inside of the array.

* Hw3 Solution

This commit will add a file in the submission folder under my name. I add answers with descriptions on big-o and persistence files.

I learned .reduce() function by doing the persistence assignment.